Tests/requirements.txt
# SuperClaude MCP Test Dependencies

# HTTP client for API testing
requests>=2.31.0

# WebSocket client for Sequential testing
websocket-client>=1.6.4

# Testing utilities
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-timeout>=2.2.0

# Performance testing
locust>=2.17.0

# Mocking for unit tests
pytest-mock>=3.12.0
responses>=0.24.1

# Code coverage
pytest-cov>=4.1.0

# Concurrent testing
aiohttp>=3.9.1
httpx>=0.25.2